Application/StudyHub.Parser/requirements.txt
fastapi
uvicorn
playwright
camelot-py[cv]
pandas
opencv-python-headless